bugfix: Fix app manifest. (#118)
* bugfix: Fix app manifest.

This can sometimes get out of sync if folks are not using the `make
apps` command or if merge conflicts are resolved through the UI. This
commit cleans up this file.

* bugfix: Fix moonphase app.

This commit fixes an issue where moonPhase or phaseImage might not be
defined by setting them to the first phase before attempting to find the
correct phase. There is likely a better way to do this but I wanted to
do the minimal changeset to get the tests to pass.