bigclock: Set defaults for schema. (#150)
This commit adds defaults for schema toggless. The issue is, we don't know where
the toggle should be by default if these are not set. For most toggles,
the default is False, and config.get("foo", False) typically defaults to
False. But in this case, the default is False but the config value
defaults to True in some cases.