Duolingo: Add schema option to display Today's XP score in the chart header. (#337)
* Push Duolingo app 🦉

* A few cosmetic tweaks

* Lighten grey shade on progress bar to make it more visible

* Add feature to hide applet from display if no lessons have been completed in the last week

If the two week display is being used, this is hidden if no lessons have been done for two weeks

* Add animated completed progress bars

* Fix progress bar multiplier

* Add Readme and screenshots

* Minor fix to owl picker

* Improved owl animations

- Reworked Flying and Angry Duo!
- Added Dancing Duo!

* Linty goodness

* Remove destiny2stats from apps.go to try and fix error

* Fix to declare hide_duolingo_in_rotation variable outside of it statement

* Fix for when no XP data is returned for a valid user

This can happen because the user has not completed a lesson in the last 7 days.

* Re-lint!

* Tweak: Add a more detailed app description

* Fix for error that occurs when chart is displayed, but there is no XP target specified

* Fix spacing on Error screen

* Lint

* Add fading progress indicators for today on the one week chart

* Add fading progress indicators for today on the two week chart

* Forgot to lint!

* Fix owl picker on chart view

* Make chart progress fader visible if no lessons have been done today yet

* On two week chart, change all of last weeks bars to dark grey

This was kind of cluttered before since, for last week, also it showed a diferent colored bar based on whether the goal was hit or not.  I decided to clean up the display to make this weeks data more prominent. Last week's data is consistently just displayed as a dark grey bar, regardless of whther or not the XP goal was achieved.

* Lint again!

* Fix so applet defaults work again

If you want to self host, you need these working. It does not affect the community version. I also added an option to display Today's XP score in the chart header.

* Added a few additional XP targets

The highest XP target is now 500 which should keep even the most dedicated student busy!