ESPN News: Fixed Vertical scroll (#369)
* ESPN News: Fixed Vertical scroll

So vertical scroll was initially having issues with cutting off headlines, so made it so a max of 12 lines can be seen in each headline and all headline has an equal width of space, so there will be a variable width between headlines some large, but figure it is better to be consistent if anything.

Also removed the fail and will return an error message if data cannot be obtained

* Fix vertical issue again.

Readjusted to fix vertical issue by submittin a PR fix to wrapped text in pixlet.

Without that merge, the last line will just be cut off, so won't be much different for vertical scroll prior to this PR

* removed base64 module as it is not used

* Adjusted marquee offset start and end

marquee offset and start set to 32 for vertical and 64 for horizontal. should help with reading the first word or two when scrolling. vertical still jumps when cycling for some reason